Crate sxpres

Structs

Cli

Cli from Clap.

Element

An slide is built from elements that are rendered to a <div class=element> that respect the SxPres philosophy.

HTML

The final HTML.

Slide

Each fully presentable slide from the entire slideshow. It is a <div class=slide> that will be formatted by CSS to fill the screen and respect the Javascript controls.

Enums

ElementNature

Define the nature of the elements. This can help organize the way multiple elements will be arranged. The order of declaration matters since that organize() uses sort_by_key() method.

Constants

COMMENT_MARKER

SEPARATOR

STD_OUTPUT_FMT

TAG_DRAFT

TAG_FOOTER

TAG_HEADING

TAG_IMAGE

TAG_LOGO

TAG_MARKER

TAG_MERMAID

TAG_MERMAIDSCRIPT

TAG_ORDLIST

TAG_SUBHEADING

TAG_TABLE

TAG_TEXT

TAG_ULIST

TAG_VIDEO

Traits

CleanTag

IsComment

Recognise a comment line.

Organize

SplitOnTag

Split a Vec<String> right on a tag, grouping them into a new Vec<String>. The result of this process is a Vec<Vec<String>>, a primite form of slide.

Functions

file_base64

Convert external files into raw base64 data to be embedded into the final HTML.

footer

Render a <footer> element with a foot message on all slides.

heading

The <h1> rendering function.

image

The <image> rendering function. Captions of figures are buided from the third line fowards on the raw_element.

input

Treats all the forms of input using Clap.

is_element_ok

Do all the checks necessary to validate a raw_element as Element.

logo

Render a <image class=logo> that is treated in CSS as a fixed right top logomark to the slides.

mermaid

The <pre class=mermaid> element rendering function. Resolves a mermaid diagram passed line-by-line.

ordlist

Generate an <ol> style listing.

output

Uses Clap to handle the output.

render

Finally condense back a Vec<Slide> into HTML that can be printed or outputed.

subheading

The <h2> rendering function.

table

The <table> rendering function.

text

The <p> rendering function.

ulist

The <ul> rendering function.

video

The <video> rendering function.